refactor(store): split orphan-file cleanup into RecordingStore.OrphanCleanup.cs
